refactor: move all props helpers into utils
